Almost there...

+++ Little Family Tree/data/LocalRelationship.swift
@@ -1,6 +1,6 @@
 class LocalRelationship {
-	var id : Int!
-	var id1 : Int!
-	var id2 : Int!
+	var id : Int64!
+	var id1 : Int64!
+	var id2 : Int64!
 	var type : RelationshipType!
 }
