+++ Little Family Tree/data/Media.swift
@@ -1,7 +1,7 @@
 import Foundation
 
 class Media {
-	var id : Int!
+	var id : Int64!
 	var familySearchId : NSString?
 	var localPath : NSString?
 	var type : NSString?
